+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-05-13 12:11:48
-// Commit: feat(network): add async network stream provider
+// Updated at: 2026-05-13 13:15:07
+// Commit: chore(network): update dependencies and cleanup unused imports in network
 
 class networkComponentState {
-  final String commitRef = '1098';
-  final String lastMessage = 'feat(network): add async network stream provider';
-  final DateTime timestamp = DateTime.parse('2026-05-13T12:11:48+05:00');
+  final String commitRef = '1101';
+  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
+  final DateTime timestamp = DateTime.parse('2026-05-13T13:15:07+05:00');
 }